service_provider/pyproject.toml
[tool.poetry]
name = "2FA_service_provider"
version = "0.1.0"
description = ""
authors = ["markdauhashei"]
readme = "README.md"
packages = [{include = "service_provider"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-dateutil = "^2.8.2"
asyncio = "^3.4.3"
pytest = "^7.2.1"
aiohttp = "^3.8.3"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.3"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
